chore: add '@ui5/theming-ngx'  dependency always (#134)
Prior to this PR the `@ui5/theming-ngx` package was missing from the projects dependencies. 

**Issue**
Prior to this PR, 
-  If you get started from scratch with
`ng create my-app`
- And if you install @ui5/webcomponents-ngx with
`npm install @ui5/webcomponents-ngx`,
- And then use some of the components in the app and run the app
`ng serve`
- you will get an error that an import to `@ui5/theming-ngx` could not be resolved
and you need to manually install the `@ui5/theming-ngx` package in addition no matter you want or don't need theming

Most probably it was optimisation effort to exclude the package and only add the dependency if the app requires theming. However, it's not working fine. Even if we use `ng add @ui5/webcomponents-ngx` the user is not asked to install it and the result is the same - we will get an error on `ng serve`.

**Solution**
Add `@ui5/theming-ngx` as dependency unconditionally to avoid breaking on first steps of the usage.

Fixes: https://github.com/SAP/ui5-webcomponents-ngx/issues/62